LF-3299-2 regulator: add regulator driver support for FP9931
The FP9931 includes different kinds of regulators: LDO, Boost
Converter, Buck-Boost Converter, Charge Pump and VCOM Buffer.
The typical voltages power on sequence is:

	V3P3 -> VGL -> VNEG -> VGH -> VPOS -> VCOM

The DISPLAY regulator is a virtual one which behaves like a
switch to trigger the voltages power on or power down sequence
except V3P3, since it must be switched on or off manually.

Signed-off-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Reviewed-by: G.n. Zhou <guoniu.zhou@nxp.com>
[squash fix warning: 'static' is not at beginning of declaration ]
Signed-off-by: Robby Cai <robby.cai@nxp.com>
[squash Defer probe when retrieve gpio which is not ready]
Signed-off-by: Robby Cai <robby.cai@nxp.com>