LF-5320: drm/bridge: nxp-seiko-43wvfig: fix boot freeze
The following:

commit b776b0f00f24 ("drm: mxsfb: Use bus_format from the nearest bridge if present")

added a check in mxsfb driver that retrieved the bus_format from the
bridge itself. However, the nxp-seiko-43wvfig driver did not provide a
hook for retrieving the input formats and the boot process hung with a
NULL pointer exception since there was no bridge state to retrieve.

The fix is to add the atomic_get_input_bus_fmts() hook, together with
the required duplicate_state(), destroy_state() and reset(), which are
necessary when any atomic hook is implemented.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Tested-by: Robert Chiras <robert.chiras@nxp.com>