LF-7114-1: arm64: dts: imx8mp: support 1XL WiFi module on imx8mp
According to 88W9098 datasheet, GPIO[15] and GPIO[16] need to be set to
11 to select the default Reference Clock Frequency. The muRata 1XL
module connect the 88w9098 GPIO[15] and GPIO[16] to M.2 pin20 & pin21,
which means the M.2 pin20 & pin21 need to be set to 11 when bring up the
1XL wifi module.

But the pin21 on imx8mp is low level by default, which is not align with
the M.2 specification that this pin need to be set as input high, also
may cause the 1XL wifi/BT cannot work, so need to config this pin to
input high to meets the M.2 specification and support the 1XL module.

Also remove the pinctl in imx8mp-evk-usdhc1-m2.dts to avoid the pinctl
conflict.

Signed-off-by: Sherry Sun <sherry.sun@nxp.com>
Reviewed-by: Haibo Chen <haibo.chen@nxp.com>