LF-7705-1 Revert "LF-562 arm: dts: add the vgen3 into pcie node"
This reverts commit afc6aefd1b4913fab8650a34c6e01641c5c32088.

The epdev_on regulator is not needed anymore, revert it here.
Because that the vgen3 had been set in the following commit.
commit 9baabac707a5 ("ARM: dts: imx6qp-sabresd: Enable PCIe support")

And the epdev_on regulator is firstly introduced by
commit cfe220b2f993 ("MLK-20685-05 arm64: dts: imx8qm/qxp: add cyw4356 and cyw4359 support").

Because cyw4356 and cyw4359 are not supported anymore, this commit is
reverted. Clean up the epdev_on regulator here too.

Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>
Reviewed-by: Sherry Sun <sherry.sun@nxp.com>