IMXBSP-101 net: stmmac: fix wrong place to call pinctrl state change
The previous patch put the phylink_suspend after the
pm_runtime_force_suspend to fix an issue. But in phylink_suspend, it
will call mdio read/write. If the iomux of the pins of MDIO bus have
been changed to others, the MDIO bus will not functional.
So put pinctrl_pm_select_sleep_state to the right place to fix this issue.

Fixes: 90702dcd19c0 ("net: stmmac: fix MAC not working when system resume back with WoL active")
Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>