LF-10970 neutron: add power_state flag to improve pm resume flows.
If there are no inference jobs, the neutron core is not started by default.
Therefore, if the neutron core was stopped before sleeping, the neutron core should not be started in resume() function.

This patch adds the power_state flag to determine whether the neutron core needs to be started in the resume() function.

Signed-off-by: Jiwei Fu <jiwei.fu@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>