LF-13220: media: ov5640: Fix camera the frist few frames are green image
When start the capture pipeline, we found 1st few frames are intermittent
green, but after few frames stable time, the green images will be
disappear. The root cause is that it needs software reset and software
power down before the ov5640 initial sequences are configured. Meanwhile
if set the software reset or software power down, it will halt the ov5640
internal clock or start the internal clock, it need 10~15ms stable time.

Signed-off-by: Alice Yuan <alice.yuan@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>