net: bridge: switchdev: guard br_switchdev_vlan_replay() with CONFIG_BRIDGE_VLAN_FILTERING
br_switchdev_vlan_replay() does nothing meaningful when
CONFIG_BRIDGE_VLAN_FILTERING is disabled. It gets by, because
br_vlan_group(), nbp_vlan_group() and co. have shim definitions, and the
code falls through when VLAN filtering is compiled out.

In a future patch, we will want to dereference br->vlan_proto, which is
only available under CONFIG_BRIDGE_VLAN_FILTERING. We could add an
accessor for that (but we'd have to touch all code that dereferences
it), or we could just make the entire body of br_switchdev_vlan_replay()
be conditionally compiled depending on CONFIG_BRIDGE_VLAN_FILTERING.

It is more convenient to opt for the latter.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>