net: sdk_dpaa: fix more build errors
Some are due to upstream work such as the include/linux/of_platform.h
header split, or the addition of netmem to skb_frag_t which revealed
that we should have gone through the skb_frag_size() accessor in the
first place. Then there are the bugs which have been there since
forever, like for example the fact that CEETM performs floating point
arithmetic in the kernel, something which is not valid.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>