LF-14260-13 net: dsa: netc: add buffer pool support
The buffer pool is a quantity of memory available for buffering a group
of flows (e.g. frames having the same priority, frames received from the
same port), while waiting to be transmitted on a port. The buffer pool
tracks internal memory consumption with upper bound limits and optionally
a non-shared portion when associated with a shared buffer pool. Currently
the shared buffer pool has not been associated with the buffer pool. We
will add the shared buffer pool support in the future. In addition, the
flow control feature of the switch is also implemented through the buffer
pool.

Based on the NETC block guides of other SoCs, the number of ports, number
of buffer pools, and the shared memory buffer sizes vary. Therefore, the
initialization of buffer pool table is better tailored to the platforms.
Hence the bpt_init() is added to netc_switch_info so that each platform
could implement its own callback. Furthermore, the flow control of each
port is also implemented based on the buffer pool, so the
port_tx_pause_config() is also added to struct netc_switch_info.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>