drm: bridge: cdns-hdmi-cec: Always initialize CEC driver
Previously, the CEC driver initialization was conditionally enabled
based on the CONFIG_DRM_CDNS_HDMI_CEC configuration.
This could lead to the CEC driver not being initialized
when built as a module, even if the module was loaded.

This patch removes the conditional check and ensures
that the CEC driver is always initialized when the HDMI driver is probed.

This fixes the issue and eliminates the build warnings
related to missing function prototypes.
../drivers/gpu/drm/bridge/cadence/cdns-mhdp-cec.c:325:5: warning: no previous prototype for ‘cdns_mhdp_register_cec_driver’ [-Wmissing-prototypes]
  325 | int cdns_mhdp_register_cec_driver(struct cdns_mhdp_cec *cec)
      |     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../drivers/gpu/drm/bridge/cadence/cdns-mhdp-cec.c:353:5: warning: no previous prototype for ‘cdns_mhdp_unregister_cec_driver’ [-Wmissing-prototypes]
  353 | int cdns_mhdp_unregister_cec_driver(struct cdns_mhdp_cec *cec)

Signed-off-by: Sandor Yu <Sandor.yu@nxp.com>