net: sdk_fman: remove dTSEC and TGEC code
These MACs are only used on PowerPC, which is not supported in Linux
Factory. The LS1043A and LS1046A contain mEMAC.

Also rename the improperly called DTSEC_SUPPORTED from generic code to
what is now called MEMAC_SUPPORTED, because now only MEMAC is the only
MAC type there is.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>