LF-14222-2 media/i2c: ox03c10: introduce standard controls
This adds standard controls for exposure, analogue gain and digital gain
which will temporarily work in parallel with the custom ones.
Eventually, the custom controls will be removed and only the standard
ones will be used.

The new controls, however, needs a complicated distribution function
that takes all 3 user provided values of exposure, analogue gain and
digital gain as inputs and computes the proper values that will be
written to the sensor registers.

In order to make this process a little more efficient and avoid writing
the same registers 3 times, add a caching mechanism that will avoid
writing a register if its value didn't change.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Mirela Rabulea <mirela.rabulea@nxp.com>