net: dpaa2-mac: reorganize code around gathering statistics
In a recent commit, how the MAC statistics are gathered was updated so
that the driver can directly request all counters through a single
firmware call. The code as it stood before this patch was not easily
extendable to cover other groups of counters as well.

This patch transforms the dpaa2_mac_setup_stats() and
dpaa2_mac_clear_stats() functions into generic ones which take a
'struct dpaa2_mac_stats' as parameter in order to know where to store
the newly allocated DMA memory. The dpmac_counter structure is also
introduced so that the setup code can use the dpmac_counter_id when
creating the list of requested counters.

Since we are planning to add rmon statistics, rename the
DPAA2_MAC_NUM_STATS into DPAA2_MAC_NUM_ETHTOOL_STATS to make it
easier to extend the support to other types of statistics.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>