LF-14278-1: arm64: dts: imx8mm: fix ov5640 camera met stream on failure
When do capture stream on/off more than 1000 tims, it will meet the
stream on failed issues.

[ 3991.601557] imx7-csi 32e20000.csi: pipeline start failed with -22
                VIDIOC_STREAMON returned -1 (Invalid argument)

The reason is when start stream, the driver will call pm_runtime_resume
function where the sensor will be re-initialized but failed.
The timing sequence for combination of PWDN and RST pin seems not to make
it initialized successfully all the time. Use only PWDN pin to stabilize
the initialization. Run over 40000 times without failure.

Fixes: '99e5d6d2741e ("arm64: dts: imx8mm-evk: Add camera support")'
Signed-off-by: Alice Yuan <alice.yuan@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>