LF-14482: neutron: request uncached neutron firmware
There are temporary firmware files that could be generated when run a
NPU case, and deleted after case finished.

The original function request_firmware() assigns firmware name to the
device. When the neutron device enters the suspended state, the firmware
will be cached per request_firmware() design. If the firmware has been
deleted, the linux kernel will report error: "failed with error -2".

In fact, the temporary firmwares don't need to be cached, so we use the
request_firmware_into_buf() function to load the firmware to avoid caching
when system enter suspend.

This patch is used to fix the following error message in the
power-management suspend test.

[ 2593.182316] PM: suspend entry (deep)
[ 2593.193843] Filesystems sync: 0.007 seconds
[ 2593.198991] (NULL device *): Direct firmware load for NeutronFw87e790501462 failed with error -2
[ 2593.198995] (NULL device *): Direct firmware load for NeutronFw87e790509029 failed with error -2
[ 2593.198993] (NULL device *): Direct firmware load for NeutronFw87e790504638 failed with error -2
[ 2593.199012] (NULL device *): Falling back to sysfs fallback for: NeutronFw87e790501462
[ 2593.199012] (NULL device *): Falling back to sysfs fallback for: NeutronFw87e790504638
[ 2593.318160] Freezing user space processes
[ 2593.324160] Freezing user space processes completed (elapsed 0.001 seconds)
[ 2593.331190] OOM killer disabled.

Signed-off-by: Jiwei.Fu <jiwei.fu@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>