LF-12736-3 nvmem: imx93: update i.MX93 fusemap information
This update is according to EdgeLock Secure Enclave User Guide i.MX 93
A1 external FW v2.0.0 64d8ef47 (0db11).

When the value of OSCCA_FUSE_READ_DIS is 1, the chip is not allowed to
access the OCOTP registers. Fuses can be read through an API call to the
ELE, so change to use the ELE Baseline API interface.

Signed-off-by: Alice Guo <alice.guo@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>