LF-12736-6 nvmem: imx93: support to get mac address by an nvmem provider
This patch is used to support to get the MAC address by an nvmem
provider.

The mac address stored in fuse is big endian. Only after the converted
mac address will be correct and it can be provided to the ethernet
framework.

squash e1c36093dc6df TBS: nvmem: imx93: avoid to free a uninitialized pointer

Signed-off-by: Alice Guo <alice.guo@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>