LF-12736-7 nvmem: imx95: add the support to return the mac addr of PFs
The 3 PFs of imx95 use 9 mac addresses, but only saves one mac base
address in fuse, which is the mac address of the first PF. The others
PF mac addresses need to be calculated through offsets.
According to the ocotp framework, adding the PF offset to bits 15~12 of
offset. This will not affect the dts node setting of other fuse devices.
In imx95, the value range of offset is 0~0x984.

squash 738acc57b5b99 TBS: nvmem: imx95: refine to reading the i.MX95 MAC address

Signed-off-by: Alice Guo <alice.guo@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>