net: enetc: add .set_preemptible_tcs() to struct enetc_pf_ops
The subsequent patch will add MAC merge layer support for i.MX95 ENETC
PF, but the registers of i.MX95 and LS1028A are different in configuring
preemptible TCs. These differences are not only the register offsets,
but also the register fields. Therefore, in order to be compatible with
different platforms, .set_preemptible_tcs() hook is added to struct
enetc_pf_ops.

Signed-off-by: Wei Fang <wei.fang@nxp.com>