net: enetc: add time specific departure support for i.MX95 ENETC PF
ENETC implements time specific departure capability, which enables the
user to specify when a frame can be transmitted. When this capability
is enabled, the device will delay the transmission of the frame so that
it can be transmitted at the precisely specified time.

The departure time of a frame is specified in the transmit BD[TX_START]
field. The value of this field is expressed as bit [29:5] of the 1588
synchronized nanosecond time. Since bit 31 and 30 are always 0, the time
window is approximately 1 second and then wraps around. As a result, a
departure time of up to approximately 0.5 seconds in the future from the
current time, may be specified so that when hardware schedules the frame,
it can determine whether the departure time is in the past or in the
future. Note that the unit is 32 ns, since bit 0 to 4 are not included.
To indicate the presence of a departure time in the transmit BD, the
BD[FL] must be set to 10b. This part of the implementation is the same
as LS1028A, so i.MX95 can fully reuse this part of the code.

Example for time specific departure with hardware offload:

tc qdisc add dev eth0 parent root handle 100: mqprio num_tc 6 \
map 0 0 1 1 2 3 4 5 queues 1@0 1@1 1@2 1@3 1@4 1@5 hw 1

tc qdisc replace dev eth0 parent 100:6 etf clockid CLOCK_TAI \
delta 30000 offload

Signed-off-by: Wei Fang <wei.fang@nxp.com>