net: enetc: add rate policer table support
The Rate Policer table contains the rate policer configuration and
operational information for a NETC function. There is one Rate Policer
table for each ENETC instance. Each table entry contains a set of
parameters that defines a single rate policer instance. A rate policer
instance controls the amount of traffic that is allowed to go through
for a particular received flow of frames.

Signed-off-by: Wei Fang <wei.fang@nxp.com>