net: enetc: add ingress stream identifier table support
The Ingress Stream Identifier table contains a set of entries each
capable of classifying incoming traffic using a mix of L2 parsed and
arbitrary data fields, into streams. As a result of an entry match,
an Ingress Stream Entry ID is provided, which is a locally significant
identifier (global to the switch), that identifies a stream within an
ENETC instance or the switch.

Signed-off-by: Wei Fang <wei.fang@nxp.com>