net: enetc: add ingress port filter table support
The Ingress Port Filter table contains a set of filters each capable of
classifying incoming traffic using a mix of L2, L3, and L4 parsed and
arbitrary field data. As a result of a filter match, several actions can
be specified such as on whether to deny or allow a frame, overriding
internal QoS attributes associated with the frame and setting parameters
for the subsequent frame processing functions, such as policing, ingress
mirroring, stream identification. Each entry corresponds to a filter.

Signed-off-by: Wei Fang <wei.fang@nxp.com>