net: enetc: move SRIOV related interfaces to enetc-pf-common
i.MX95 ENETC also supports VF and the mailbox mechanism between PF and
VF. These functions are basically the same as those of LS1028A ENETC,
so only minor modifications need to be made to the implementation of
LS1028A for reuse. Therefore, these common interfaces are moved to the
enetc-pf-common driver.

Signed-off-by: Wei Fang <wei.fang@nxp.com>