net: netc-blk-ctrl: add preparation work for suspend/resume and WoL support
In previous hardware implementations, NETCMIX does not power off during
suspend. If NETCMIX powers off, all NETCMIX devices need to be
reinitialized in the resume process. So, add this function support.

When using WOL mode, NETCMIX will not power off during suspend.

The main responsibility of the BLK_CTRL driver is to manage the ipg_clk
on suspend/resume, and to offer API to consumer drivers, such as NETCv4
PSI, switch and PTP timer module, to control and query the WoL status.

Reviewed-by: Claudiu Manoil <claudiu.manoil@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>