LF-11321-2 net: enetc: make PF driver only provide SRIOV interface
The ENETC PF driver provides SRIOV interface to enable its own VFs. But
there is an use case that ENETC PF is owned by M core and VFs are owned
by A core (Linux). In this case the ENETC PF driver of Linux should only
provide SRIOV interface so that we can enable VFs through the standard
interface and command. For .probe() and .remove() success is returned
directly. The below command is used to enable the ENETC VFs.

echo 2 > /sys/bus/pci/devices/0002\:00\:10.0/sriov_numvfs

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Reviewed-by: Frank Li <frank.li@nxp.com>