LF-14557 dmaengine: fsl-edma: modify the channel parameter config method
For example, the audio micfil uses AHB edma channel6 on imx95. In order to
find the channel6 matched, need to traverse channel0-5. In this process,
the channel0-5 parameter are modified if the channel0-5 isn't requested.
If test memcpy on channel0-5 using dmatest, it can fail. Because the
channel parameter config is wrong.

This issue only impact edma memcpy test when the channel is fixed.

This patch configs the request channel only if have found the channel.

Signed-off-by: Joy Zou <joy.zou@nxp.com>
Reviewed-by: Frank Li <Frank.Li@nxp.com>