LF-14779-5 drm/imx: dcif: Remove clock mux hardcoding
Displaymix CSR registers were set in the runtime resume routines,
hardcoding the clock mux parent to lvds_pll_div7 making any setting in
DT useless. This was probably to work around a BLK CTL issue that reset
the registers because the power domain supply was cut off when runtime
suspend was called. We don't need this now and that code can be safely
removed.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Acked-by: Liu Ying <victor.liu@nxp.com>