LF-15096 net: dsa: netc: add EEE support
Each user port of the NETC switch supports transmitting LPI sequence to
remote partner. The PMa_SLEEP_TIMER can be configured to automatically
transmit the LPI sequence once the MAC transmitter has been idle for the
configured amount of time. When this dynamic LPI mode is used, the MAC
will also automatically stop transmitting the LPI sequence whenever a
new frame is available for the MAC to transmit.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>