LF-15142 drm/imx: dpu95: Add VScaler support
Now that VScaler4 issue in i.MX95 A0/A1 SoC is fixed in i.MX95 B0 SoC,
enable VScaler support for B0.  Only add scaling support for VScaler4
in KMS because it's not clear how to support de-interlacing.  While at
it, initialize VScaler9 as a blit engine component in the DPU95 core
driver.

Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>