LF-15091-1 media: i2c: ox03c10: Sync custom controls with standard ones
When standard controls for exposure, analog/digital gains are used to
change sensor settings, the custom controls remained with their old
values. This is misleading as they need to reflect the current sensor
setting. Fix this by creating helper functions for activating and
launching the group hold registers and call the
__v4l2_ctrl_s_ctrl_compound() instead of calling the respective set
functions directly.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>