LF-15200: mxc: vpu: wave6: Fix an rgb encoding failure issue
After update to driver v1.3.12, the encoding will fail, when the source
format is rgb.

The reason is the type coefficient variables is u32, but in
wave6_set_csc(), some negative value is set.  but in
wave6_check_enc_param(), it will return error if the u32 value is great
than 1023. then the negative u32 coefficient will always fail the
wave6_check_enc_param().

The valid value range is [-512, 511], in wave6_check_enc_param(),
convert the coefficient to a positive value, then check its range.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Zhou Peng <eagle.zhou@nxp.com>