LF-425-2 drm/bridge: sec-dsim: use 'attach-bridge' for bridge attach
Since the port nodes under dsim device can be detached
dynamically by other devices. So use the 'attach-bridge'
property to imply that if there is any available bridge
endpoint node exists. If not, the panel node existence
is required to be detected. If still not, whole display
components binding should be ended which can avoid the
endless defer probe problem when no available display
devices can be attached to dsim.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
[ Liu Ying: Fix conflict to rebase on next-20250417 ]
Signed-off-by: Liu Ying <victor.liu@nxp.com>