net: enetc: VFs are not allowed to set the RSS key
The VFs of ENETC do not have privilege to set the RSS key, because the
key registers are owned by PF. Currently, there is no check on setting
the RSS key for VF, users may mistakenly believe that VF supports setting
the key. Therefore, add a check for this situation.

Fixes: d382563f541b ("enetc: Add RFS and RSS support")
Signed-off-by: Wei Fang <wei.fang@nxp.com>