LF-15623-9 net: dsa: netc: restore tc-taprio when switch resumes
If the NETC is powered off in suspend mode, the tc-taprio configuration
will be lost. Therefore, when the system resumes, we need to restore
the tc-taprio configuration to ensure that tc-taprio can still resume
working after the system resumes.

Note that the current patch only supports restoring the last set gate
control list, while the time gate sheduling table (TGST) has two lists,
an administrative gate control list and an operational gate control
list. Therefore, if both lists are valid, only the administrative gate
control list can be restored. So the current patch can be further
improved, but this will undoubtedly make the driver more complicated,
and perhaps it will be improved in the future.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Vladimir Oltean <vladimir.oltean@nxp.com>