LF-15623-11 net: dsa: netc: restore tc-flower tables when switch resumes
If NETC is powered off when system suspends, all the configurations of
the tc-flower tables will be lost. Therefore, it is necessary to restore
them when the system resumes, to ensure that the tc-flower still works
properly.

In addition, since multiple rules may share the same gate/police/key
table entries, so the restored flag is added to indicate the entry has
been restored, no need to set it up again.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Acked-by: Clark Wang <xiaoning.wang@nxp.com>