LF-15875-1 arm64: dts: imx943: add ENETC sharing use case support
For the ENETC sharing use case of i.MX943, all three VFs of ENETC 3 are
owned by A-core, and all other resources of the whole NETC are owned by
M-core, including the entire PCIe ECAM space of NETC. Therefore, A-core
cannot directly access the PCIe ECAM space, but the Linux PCIe driver
needs to perform enumeration to discover the PCIe functions and then
probe the VF driver. So the M-core create a 1MB shared memory as the
virtual PCIe ECAM space, so that Linux can directly read this virtual
ECAM and perform PCIe enumeration.

In addition, since the I2C3 and I2C6 are used to reset the PHYs, so
these two buses are also controlled by M-core, so just simply disable
them in the DTS for this sharing use case. All devices which depend on
these two buses will be unavailabe on Linux side.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>