LF-15788 spi: spi-nxp-xspi: adjust the watermark setting for no-4byte data length alignment case
According to the RM, for TBDR register, a write transaction on the
flash memory with data size of less than 32 bits leads to the removal
of one data entry from the TX buffer. The valid bits are used and the
rest of the bits are discarded.

But for data size large than 32 bits, according to test, for no-4bytes
alignment data, the last 1~3 bytes will lost, because TX buffer use 4
bytes entries. Confirmed with IP team, for IP write, software need to
define the IDATSZ in alignment with the data programming of TBDR
(which is 4B aligned) register.

To cover the no-4byte data length alignment case:
1. adjust the watermark setting
2. adjust the transfer data length to make it 4bytes alignment.
   then will meet the upper watermark setting, trigger the 4bytes
   entries all pop out.
3. use extra 0xff for bytes append.

Signed-off-by: Haibo Chen <haibo.chen@nxp.com>