LF-15911: mxc: vpu: hantro_v4l2: Add support for non-coherent buffer
Set allow_cache_hints to 1 for both output and capture queue. when using
mmap type v4l2 buffer, user can choose non-coherent flag in requesting
or creating buffers, then it can help speed up the reading and writing
performance. And the videobuf2 has helped handle the synchronization.

For the daemon, it still mmap all memory in coherent, then this will not
affect the encoding and decoding.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Zhou Peng <eagle.zhou@nxp.com>