LF-16077: net: dsa: use forward tag to fill netc dsa tag on hsr port
When the netc switch port is set as hsr port. The HSR tag add and
packets duplicate will offload on hardware. The Ingress Sequence Table
of hardware will add HSR tag and FDB table will duplicate the packets to
two HSR ports.

When packets send from CPU, the packets will be added a DSA switch tag
with NETC_TAG_TO_PORT, which will bypass the ingress process and send
packets directly to egress and bypass the FDB table. We don't want this
happen when the port is set as hsr port. So we need to use
NETC_TAG_FORWARD in DSA switch tag if the port is configured as HSR
forwarding. All packets sent to one of HSR port will forward to switch,
add HSR tag on ingress cpu port and duplicate by FDB to two HSR ports.

Signed-off-by: Xiaoliang Yang <xiaoliang.yang_1@nxp.com>