phy: lynx-28g: implement phy_get_status() operations
Allow consumers of the SerDes lane to figure out the PCS and AN/LT block
address without hardcoding it in the device tree or in the driver code,
which is what they do currently.

Also, implement CDR lock status, to gain a sense of when the link is
down for 25GbE where the PCS does not implement the power down
operation.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>