MLK-26310: phy: freescale: phy-fsl-imx8-mipi-dphy: HS prepare calculation has underflow
The unsigned variable "n" underflows in the high speed data prepare
calculation. Changing this to a signed integer prevents the underflow and
allows the range check to clamp the register value at 1.

Here is the failing calculation:

      /* hs_zero: formula from NXP BSP */
      n = (144 * (dphy_opts->hs_clk_rate / 1000000) - 47500) / 10000;
      cfg->m_prg_hs_zero = n < 1 ? 1 : n;

If hs_clk_rate is the minimum value 80000000 Hz, we get -4 for n. Any
value below +1 needs to clamped at +1 per i.MX8ULP Reference Manual.

Signed-off-by: Oliver F. Brown <oliver.brown@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>