LF-16454 media/i2c: ox03c10: Fix vertical/horizontal blanking controls
Currently, when switching from 1920x1280 mode to 1920x1080, the vertical
blanking control value remains the same, instead of reflecting the new
vertical blanking time for the new mode.

The problem happened because the ox03c10_s_ctrl() callback did not
handle the HBLANK and VBLANK controls and, when the __v4l2_ctrl_s_ctrl()
was called with the new values after the mode was changed, it returned
-EINVAL, preserving the current value.

Add the V4L2_CID_VBLANK and V4L2_CID_HBLANK handling to
ox03c10_s_ctrl().

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reported-by: Celine Laurencin <celine.laurencin@nxp.com>
Tested-by: Celine Laurencin <celine.laurencin@nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Mirela Rabulea <mirela.rabulea@nxp.com>