LF-16458 arm64: dts: imx943-evk-sdwifi: set SION for usdhc3 cmd pad
Since usdhc3 data transfer will have CRC errors if usdhc2 is
reading/writing, the root cause is identified but the Errata has not
come out yet. The temporary workaround suggested by design team is to
set SION for usdhc3 cmd pad, this patch has passed the stress tests.

The following is the detailed info from Soc design team:
The cmd_en_inv (used as pad input enable signal) of usdhc3 is connected
to inverted of cmd_en of usdhc2. By default the input is enabled, so
when usdhc2 is not working, usdhc3 can still get input cmd response data
as expected. But when usdhc2 is working, it may go low, thus wrong.

Signed-off-by: Sherry Sun <sherry.sun@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>