LF-16821 drm/imx: dpu-crtc: Invalidate video modes unsupported by pixel combiner
On i.MX8QM, when the image width is larger or pixel clock is faster than
specified thresholds, the Pixel Combiner is involved. The Pixel Combiner
has constraints on the timings. These timings were verified at the CRTC
atomic_check stage, but this wasn't able to invalidate the unsupported
video modes. This patch permits to also check the Pixel Combiner
constraints in the CRTC mode_valid hook.

Signed-off-by: Thomas Antonini <thomas.antonini@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>