LF-16843 media: i2c: max96717_lib: Print serializer chip variant
It is useful, for debugging purposes, to know which variant of the
MAX96717 chip is being used, since MAX96717F supports only 3 Gbps GMSL
link speeds, while MAX96717 supports both 3 and 6 Gbps. It is possible
that, in the future, we add support for modes that require 6 Gbps link
speeds and the user needs to know which variant is being used on every
GMSL link.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Mirela Rabulea <mirela.rabulea@nxp.com>